move some versioning utils from common to worker_versioning
