+++ common/worker_versioning/worker_versioning.go
@@ -28,10 +28,13 @@ import (
 	"context"
 	"fmt"
 
+	"github.com/gogo/protobuf/types"
 	"github.com/xwb1989/sqlparser"
 	commonpb "go.temporal.io/api/common/v1"
 
 	persistencespb "go.temporal.io/server/api/persistence/v1"
+	taskqueuespb "go.temporal.io/server/api/taskqueue/v1"
+	"go.temporal.io/server/common"
 	"go.temporal.io/server/common/namespace"
 	"go.temporal.io/server/common/persistence/visibility/manager"
 	"go.temporal.io/server/common/searchattribute"
@@ -70,18 +73,14 @@ func VersionStampToBuildIdSearchAttribute(stamp *commonpb.WorkerVersionStamp) st
 // Returns -1, -1 if not found.
 func FindBuildId(versioningData *persistencespb.VersioningData, buildId string) (setIndex, indexInSet int) {
 	versionSets := versioningData.GetVersionSets()
-	setIndex = -1
-	indexInSet = -1
 	for sidx, set := range versionSets {
 		for bidx, id := range set.GetBuildIds() {
 			if buildId == id.Id {
-				setIndex = sidx
-				indexInSet = bidx
-				break
+				return sidx, bidx
 			}
 		}
 	}
-	return setIndex, indexInSet
+	return -1, -1
 }
 
 func WorkflowsExistForBuildId(ctx context.Context, visibilityManager manager.VisibilityManager, ns *namespace.Namespace, taskQueue, buildId string) (bool, error) {
@@ -98,4 +97,42 @@ func WorkflowsExistForBuildId(ctx context.Context, visibilityManager manager.Vis
 		return false, err
 	}
 	return response.Count > 0, nil
+}
+
+// StampIfUsingVersioning returns the given WorkerVersionStamp if it is using versioning,
+// otherwise returns nil.
+func StampIfUsingVersioning(stamp *commonpb.WorkerVersionStamp) *commonpb.WorkerVersionStamp {
+	if stamp.GetUseVersioning() {
+		return stamp
+	}
+	return nil
+}
+
+func MakeDirectiveForWorkflowTask(
+	stamp *commonpb.WorkerVersionStamp,
+	lastWorkflowTaskStartedEventID int64,
+) *taskqueuespb.TaskVersionDirective {
+	var directive taskqueuespb.TaskVersionDirective
+	if id := StampIfUsingVersioning(stamp).GetBuildId(); id != "" {
+		directive.Value = &taskqueuespb.TaskVersionDirective_BuildId{BuildId: id}
+	} else if lastWorkflowTaskStartedEventID == common.EmptyEventID {
+		// first workflow task
+		directive.Value = &taskqueuespb.TaskVersionDirective_UseDefault{UseDefault: &types.Empty{}}
+	}
+	// else: unversioned queue
+	return &directive
+}
+
+func MakeDirectiveForActivityTask(
+	stamp *commonpb.WorkerVersionStamp,
+	useCompatibleVersion bool,
+) *taskqueuespb.TaskVersionDirective {
+	var directive taskqueuespb.TaskVersionDirective
+	if !useCompatibleVersion {
+		directive.Value = &taskqueuespb.TaskVersionDirective_UseDefault{UseDefault: &types.Empty{}}
+	} else if id := StampIfUsingVersioning(stamp).GetBuildId(); id != "" {
+		directive.Value = &taskqueuespb.TaskVersionDirective_BuildId{BuildId: id}
+	}
+	// else: unversioned queue
+	return &directive
 }
