+++ service/history/workflow/retry.go
@@ -45,6 +45,7 @@ import (
 	"go.temporal.io/server/common"
 	"go.temporal.io/server/common/backoff"
 	"go.temporal.io/server/common/primitives/timestamp"
+	"go.temporal.io/server/common/worker_versioning"
 )
 
 // TODO treat 0 as 0, not infinite
@@ -245,7 +246,7 @@ func SetupNewWorkflowForRetryOrCron(
 	// For cron: do not propagate (always start on latest version).
 	var sourceVersionStamp *commonpb.WorkerVersionStamp
 	if initiator == enumspb.CONTINUE_AS_NEW_INITIATOR_RETRY {
-		sourceVersionStamp = common.StampIfUsingVersioning(previousMutableState.GetWorkerVersionStamp())
+		sourceVersionStamp = worker_versioning.StampIfUsingVersioning(previousMutableState.GetWorkerVersionStamp())
 	}
 
 	req := &historyservice.StartWorkflowExecutionRequest{
